Исправляет работу тестов
