+++ js/data/data.js
@@ -7,6 +7,24 @@ export const INITIAL_GAME = Object.freeze({
   answers: []
 });
 
+export const getPointsScored = (answers) => {
+  let scored = 0;
+  if (answers.length < 10) {
+    return -1;
+  }
+  answers.forEach((it) => {
+    if (!it.correct) {
+      scored -= 2;
+    } else if (it.correct && it.time < 30) {
+      scored += 2;
+    } else if (it.correct) {
+      scored += 1;
+    }
+  });
+
+  return scored;
+};
+
 
 export const outputResult = (resultAnotherPersons, currentGame) => {
   if (currentGame.timeLeft === 0) {
